* merging jmeter-extended to keptn core jmeter

* removing references to jmeter-extended

* edited some details in the dockerfile, updated go version, removed obsolete files

+++ jmeter-service/main.go
@@ -19,7 +19,7 @@ import (
 	"github.com/google/uuid"
 	"github.com/kelseyhightower/envconfig"
 
-	"github.com/keptn/go-utils/pkg/lib"
+	keptn "github.com/keptn/go-utils/pkg/lib"
 )
 
 const eventbroker = "EVENTBROKER"
@@ -56,8 +56,8 @@ func gotEvent(ctx context.Context, event cloudevents.Event) error {
 		return errors.New(errorMsg)
 	}
 
-	if data.TestStrategy == "real-user" {
-		logger.Info("Received 'real-user' test strategy, hence no tests are triggered")
+	if data.TestStrategy == TestStrategy_RealUser {
+		logger.Info("Received '" + TestStrategy_RealUser + "' test strategy, hence no tests are triggered")
 		return nil
 	}
 	go runTests(event, shkeptncontext, *data, logger)
@@ -65,72 +65,97 @@ func gotEvent(ctx context.Context, event cloudevents.Event) error {
 	return nil
 }
 
+//
+// This method executes the correct tests based on the passed testStrategy in the deployment finished event
+// The method will always try to execute a health check workload first, then execute the workload based on the passed testStrategy
+//
 func runTests(event cloudevents.Event, shkeptncontext string, data keptn.DeploymentFinishedEventData, logger *keptn.Logger) {
-	keptnHandler, err := keptn.NewKeptn(&event, keptn.KeptnOpts{})
-	if err != nil {
-		logger.Error("Could not initialize Keptn handler: " + err.Error())
-	}
+
 	testInfo := getTestInfo(data)
 	id := uuid.New().String()
 	startedAt := time.Now()
 
-	var res bool
-	res, err = runHealthCheck(data, id, logger)
-	if err != nil {
-		logger.Error(err.Error())
-		return
-	}
+	// load the workloads from JMeterConf
+	var err error
+	var jmeterconf *JMeterConf
+	jmeterconf, err = getJMeterConf(testInfo.Project, testInfo.Stage, testInfo.Service, logger)
 
-	if !res {
-		if err := sendTestsFinishedEvent(keptnHandler, event, startedAt, "fail", logger); err != nil {
+	// get the service endpoint we need to run the test against
+	var serviceUrl *url.URL
+	serviceUrl, err = getServiceURL(data)
+	if err != nil {
+		logger.Error(fmt.Sprintf("Error getting service url to run test against: %s", err.Error()))
+		if err := sendTestsFinishedEvent(shkeptncontext, event, startedAt, "fail", logger); err != nil {
 			logger.Error(fmt.Sprintf("Error sending test finished event: %s", err.Error()) + ". " + testInfo.ToString())
 		}
 		return
 	}
-	logger.Info("Health Check test passed = " + strconv.FormatBool(res) + ". " + testInfo.ToString())
 
-	var sendEvent = false
-
-	switch strings.ToLower(data.TestStrategy) {
-	case "functional":
-		res, err = runFunctionalCheck(data, id, logger)
+	// first we run a health check workload. If that fails we stop the rest
+	var healthCheckWorkload *Workload
+	var res bool
+	healthCheckWorkload, err = getWorkload(jmeterconf, TestStrategy_HealthCheck)
+	if healthCheckWorkload != nil {
+		res, err = runWorkload(serviceUrl, testInfo, id, healthCheckWorkload, logger)
 		if err != nil {
 			logger.Error(err.Error())
 			return
 		}
-		logger.Info("Functional test passed = " + strconv.FormatBool(res) + ". " + testInfo.ToString())
-		sendEvent = true
 
-	case "performance":
-		res, err = runPerformanceCheck(data, id, logger)
-		if err != nil {
-			logger.Error(err.Error())
-			return
-		}
-		logger.Info("Performance test passed = " + strconv.FormatBool(res) + ". " + testInfo.ToString())
-		sendEvent = true
-
-	case "":
-		logger.Info("No test strategy specified, hence no tests are triggered. " + testInfo.ToString())
-		sendEvent = true
-
-	default:
-		logger.Error("Unknown test strategy '" + data.TestStrategy + "'" + ". " + testInfo.ToString())
-	}
-
-	if sendEvent {
 		if !res {
-			if err := sendTestsFinishedEvent(keptnHandler, event, startedAt, "fail", logger); err != nil {
+			if err := sendTestsFinishedEvent(shkeptncontext, event, startedAt, "fail", logger); err != nil {
 				logger.Error(fmt.Sprintf("Error sending test finished event: %s", err.Error()) + ". " + testInfo.ToString())
 			}
 			return
 		}
-		if err := sendTestsFinishedEvent(keptnHandler, event, startedAt, "pass", logger); err != nil {
+		logger.Info("Health Check test passed = " + strconv.FormatBool(res) + ". " + testInfo.ToString())
+	} else {
+		logger.Info("No Health Check test workload configuration found. Skipping Health Check")
+	}
+
+	// now lets execute the workload based on the passed testStrategy
+	res = false
+	var testStrategy = strings.ToLower(data.TestStrategy)
+
+	if testStrategy == "" {
+		// no testStrategy passed at all -> we just send a successful test finished event!
+		logger.Info("No testStrategy specified therefore skipping further test execution and sending back success")
+		res = true
+	} else {
+		// lets get the workload configuration for the test strategy
+		var teststrategyWorkload *Workload
+		teststrategyWorkload, err = getWorkload(jmeterconf, testStrategy)
+		if teststrategyWorkload != nil {
+			res, err = runWorkload(serviceUrl, testInfo, id, teststrategyWorkload, logger)
+			if err != nil {
+				logger.Error(err.Error())
+				res = false
+			} else {
+				logger.Info(fmt.Sprintf("Tests for %s with status = %s.%s", testStrategy, strconv.FormatBool(res), testInfo.ToString()))
+			}
+		} else {
+			// no workload for that test strategy!!
+			res = false
+			logger.Error(fmt.Sprintf("No workload definition found for testStrategy %s", testStrategy))
+		}
+	}
+
+	// now lets send the test finished event
+	if !res {
+		if err := sendTestsFinishedEvent(shkeptncontext, event, startedAt, "fail", logger); err != nil {
 			logger.Error(fmt.Sprintf("Error sending test finished event: %s", err.Error()) + ". " + testInfo.ToString())
 		}
+		return
+	}
+
+	if err := sendTestsFinishedEvent(shkeptncontext, event, startedAt, "pass", logger); err != nil {
+		logger.Error(fmt.Sprintf("Error sending test finished event: %s", err.Error()) + ". " + testInfo.ToString())
 	}
 }
 
+//
+// Extracts relevant information from the data object
+//
 func getTestInfo(data keptn.DeploymentFinishedEventData) *TestInfo {
 	return &TestInfo{
 		Project:      data.Project,
@@ -140,6 +165,9 @@ func getTestInfo(data keptn.DeploymentFinishedEventData) *TestInfo {
 	}
 }
 
+//
+// returns the service URL that is either passed via the DeploymentURI* parameters or constructs one based on keptn naming structure
+//
 func getServiceURL(data keptn.DeploymentFinishedEventData) (*url.URL, error) {
 
 	if data.DeploymentURILocal != "" {
@@ -158,53 +186,31 @@ func getServiceURL(data keptn.DeploymentFinishedEventData) (*url.URL, error) {
 	return url.Parse(serviceURL)
 }
 
-func runHealthCheck(data keptn.DeploymentFinishedEventData, id string, logger *keptn.Logger) (bool, error) {
-	os.RemoveAll("HealthCheck_" + data.Service)
-	os.RemoveAll("HealthCheck_" + data.Service + "_result.tlf")
+//
+// executes the actual JMEter tests based on the workload configuration
+//
+func runWorkload(serviceUrl *url.URL, testInfo *TestInfo, id string, workload *Workload, logger *keptn.Logger) (bool, error) {
+
+	// for testStrategy functional we enforce a 0% error policy!
+	breakOnFunctionalIssues := workload.TestStrategy == TestStrategy_Functional
+
+	logger.Info(
+		fmt.Sprintf("Running workload testStrategy=%s, vuser=%d, loopcount=%d, thinktime=%d, funcvalidation=%t, acceptederrors=%f, avgrtvalidation=%d, script=%s",
+			workload.TestStrategy, workload.VUser, workload.LoopCount, workload.ThinkTime, breakOnFunctionalIssues, workload.AcceptedErrorRate, workload.AvgRtValidation, workload.Script))
+	if runlocal {
+		logger.Info("LOCALLY: not executiong actual tests!")
+		return true, nil
+	}
+
+	os.RemoveAll(workload.TestStrategy + "_" + testInfo.Service)
+	os.RemoveAll(workload.TestStrategy + "_" + testInfo.Service + "_result.tlf")
 	os.RemoveAll("output.txt")
 
-	testInfo := getTestInfo(data)
-	url, err := getServiceURL(data)
-	if err != nil {
-		return false, err
-	}
-	return executeJMeter(testInfo, "jmeter/basiccheck.jmx", "HealthCheck_"+data.Service, url,
-		1, 1, 250, "HealthCheck_"+id,
-		true, 0, logger)
+	return executeJMeter(testInfo, workload, workload.TestStrategy+"_"+testInfo.Service, serviceUrl, workload.TestStrategy+""+id,
+		breakOnFunctionalIssues, logger)
 }
 
-func runFunctionalCheck(data keptn.DeploymentFinishedEventData, id string, logger *keptn.Logger) (bool, error) {
-
-	os.RemoveAll("FuncCheck_" + data.Service)
-	os.RemoveAll("FuncCheck_" + data.Service + "_result.tlf")
-	os.RemoveAll("output.txt")
-
-	testInfo := getTestInfo(data)
-	url, err := getServiceURL(data)
-	if err != nil {
-		return false, err
-	}
-	return executeJMeter(testInfo, "jmeter/load.jmx", "FuncCheck_"+data.Service, url,
-		1, 1, 250, "FuncCheck_"+id, true, 0, logger)
-}
-
-func runPerformanceCheck(data keptn.DeploymentFinishedEventData, id string, logger *keptn.Logger) (bool, error) {
-
-	os.RemoveAll("PerfCheck_" + data.Service)
-	os.RemoveAll("PerfCheck_" + data.Service + "_result.tlf")
-	os.RemoveAll("output.txt")
-
-	testInfo := getTestInfo(data)
-	url, err := getServiceURL(data)
-	if err != nil {
-		return false, err
-	}
-	return executeJMeter(testInfo, "jmeter/load.jmx", "PerfCheck_"+data.Service, url,
-		10, 500, 250, "PerfCheck_"+id,
-		false, 0, logger)
-}
-
-func sendTestsFinishedEvent(keptnHandler *keptn.Keptn, incomingEvent cloudevents.Event, startedAt time.Time, result string, logger *keptn.Logger) error {
+func sendTestsFinishedEvent(shkeptncontext string, incomingEvent cloudevents.Event, startedAt time.Time, result string, logger *keptn.Logger) error {
 	source, _ := url.Parse("jmeter-service")
 	contentType := "application/json"
 
@@ -227,15 +233,19 @@ func sendTestsFinishedEvent(keptnHandler *keptn.Keptn, incomingEvent cloudevents
 			Type:        keptn.TestsFinishedEventType,
 			Source:      types.URLRef{URL: *source},
 			ContentType: &contentType,
-			Extensions:  map[string]interface{}{"shkeptncontext": keptnHandler.KeptnContext},
+			Extensions:  map[string]interface{}{"shkeptncontext": shkeptncontext},
 		}.AsV02(),
 		Data: testFinishedData,
 	}
 
-	return keptnHandler.SendCloudEvent(event)
+	return sendEvent(event)
 }
 
 func _main(args []string, env envConfig) int {
+
+	if runlocal {
+		log.Println("Running LOCALLY: env=runlocal")
+	}
 
 	ctx := context.Background()
 
@@ -252,7 +262,57 @@ func _main(args []string, env envConfig) int {
 		log.Fatalf("failed to create client, %v", err)
 	}
 
+	log.Println("Started jmeter-extended-service on ", env.Path, env.Port)
+
 	log.Fatalf("failed to start receiver: %s", c.StartReceiver(ctx, gotEvent))
 
 	return 0
+}
+
+func sendEvent(event cloudevents.Event) error {
+	if runlocal {
+		log.Println("LOCALLY: Sending Event")
+		return nil
+	}
+
+	endPoint, err := getServiceEndpoint(eventbroker)
+	if err != nil {
+		return errors.New("Failed to retrieve endpoint of eventbroker. %s" + err.Error())
+	}
+
+	if endPoint.Host == "" {
+		return errors.New("Host of eventbroker not set")
+	}
+
+	transport, err := cloudeventshttp.New(
+		cloudeventshttp.WithTarget(endPoint.String()),
+		cloudeventshttp.WithEncoding(cloudeventshttp.StructuredV02),
+	)
+	if err != nil {
+		return errors.New("Failed to create transport:" + err.Error())
+	}
+
+	c, err := client.New(transport)
+	if err != nil {
+		return errors.New("Failed to create HTTP client:" + err.Error())
+	}
+
+	if _, _, err := c.Send(context.Background(), event); err != nil {
+		return errors.New("Failed to send cloudevent:, " + err.Error())
+	}
+	return nil
+}
+
+// getServiceEndpoint gets an endpoint stored in an environment variable and sets http as default scheme
+func getServiceEndpoint(service string) (url.URL, error) {
+	url, err := url.Parse(os.Getenv(service))
+	if err != nil {
+		return *url, fmt.Errorf("Failed to retrieve value from ENVIRONMENT_VARIABLE: %s", service)
+	}
+
+	if url.Scheme == "" {
+		url.Scheme = "http"
+	}
+
+	return *url, nil
 }
